android_app/test_wiki.py
from allure import step
from appium.webdriver.common.appiumby import AppiumBy
from selene import browser


def test_open_first_search():
    with step('Type search'):
        browser.element((AppiumBy.ACCESSIBILITY_ID, "Search Wikipedia")).click()
        browser.element((AppiumBy.ID,
                         "org.wikipedia.alpha:id/search_src_text")).type('Auto')

    with step('Open first article'):
        results = browser.all((AppiumBy.ID, 'org.wikipedia.alpha:id/page_list_item_title'))
        results.first.click()