requirements.txt
pytest==8.3.3
allure-pytest
selene==2.0.0rc9
Appium-Python-Client
python-dotenv==1.0.1
requests>=2.32.1
pydantic~=1.10.9